:recycle: refactor: move files and update paths for admin
