+++ frontend/src/pages/admin/doctors/index.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import api from '../../api/api';
+import api from '../../../api/api';
 
 export default function AdminDoctors() {
     const [name, setName] = useState('');
