+++ frontend/src/pages/admin/patients/index.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import api from '../../api/api';
+import api from '../../../api/api';
 
 export default function Patients() {
     const [name, setName] = useState('');
